@@ -8,34 +8,37 @@
     "predeploy": "npm run build",
     "deploy": "gh-pages -b gh-pages -d build",
     "eject": "react-scripts eject",
+    "prepare": "husky install",
     "start": "react-scripts start",
     "test": "react-scripts test"
   },
   "dependencies": {
     "baseui": "^12.2.0",
     "glamor": "^2.20.40",
-    "react": "^16.8.0",
-    "react-bootstrap": "^2.6.0",
+    "react": "^18.2.0",
+    "react-bootstrap": "^2.9.0",
     "react-cursor-custom": "^1.2.4",
-    "react-dom": "^16.8.0",
+    "react-dom": "^18.2.0",
     "react-ga": "^3.3.1",
-    "react-icons": "^4.7.1",
+    "react-gtm-module": "^2.0.11",
+    "react-helmet-async": "^2.0.5",
+    "react-icons": "^4.11.0",
     "react-reveal": "^1.2.2",
-    "react-router-dom": "^5.2.0",
+    "react-router-dom": "^6.16.0",
     "react-scripts": "^5.0.1",
-    "styled-components": "^5.2.1",
-    "styletron-engine-atomic": "^1.4.6",
+    "styled-components": "^6.0.8",
+    "styletron-engine-atomic": "^1.5.0",
     "styletron-react": "^6.1.0",
     "webpack": "^5.91.0"
   },
   "optionalDependencies": {
-    "fsevents": "^2.3.2"
+    "fsevents": "^2.3.3"
   },
   "devDependencies": {
-    "husky": "^4.3.8",
-    "lint-staged": "^10.5.4",
-    "prettier": "2.2.1",
-    "prettier-package-json": "^2.2.0"
+    "husky": "^8.0.3",
+    "lint-staged": "^14.0.1",
+    "prettier": "3.0.3",
+    "prettier-package-json": "^2.8.0"
   },
   "browserslist": {
     "production": [
@@ -51,11 +54,6 @@
   },
   "eslintConfig": {
     "extends": "react-app"
-  },
-  "husky": {
-    "hooks": {
-      "pre-commit": "lint-staged"
-    }
   },
   "lint-staged": {
     "*.json": "prettier-package-json --write",
