@@ -1,5 +1,5 @@
 import React from "react";
-import { Route, Switch, HashRouter } from "react-router-dom";
+import { Route, Routes, HashRouter } from "react-router-dom";
 import Home from "../pages/home/HomeComponent";
 import Splash from "../pages/splash/Splash";
 import Education from "../pages/education/EducationComponent";
@@ -13,80 +13,48 @@ export default function Main(propss) {
     return (
       <div>
         <HashRouter basename="/">
-          <Switch>
+          <Routes>
             <Route
               path="/"
-              exact
-              render={(props) => (
-                <Splash
-                  {...props}
-                  theme={propss.theme}
-                  setTheme={propss.setTheme}
-                />
-              )}
+              element={
+                <Splash theme={propss.theme} setTheme={propss.setTheme} />
+              }
             />
             <Route
               path="/home"
-              render={(props) => (
-                <Home
-                  {...props}
-                  theme={propss.theme}
-                  setTheme={propss.setTheme}
-                />
-              )}
+              element={<Home theme={propss.theme} setTheme={propss.setTheme} />}
             />
             <Route
               path="/experience"
-              exact
-              render={(props) => (
-                <Experience
-                  {...props}
-                  theme={propss.theme}
-                  setTheme={propss.setTheme}
-                />
-              )}
+              element={
+                <Experience theme={propss.theme} setTheme={propss.setTheme} />
+              }
             />
             <Route
               path="/education"
-              render={(props) => (
-                <Education
-                  {...props}
-                  theme={propss.theme}
-                  setTheme={propss.setTheme}
-                />
-              )}
+              element={
+                <Education theme={propss.theme} setTheme={propss.setTheme} />
+              }
             />
             <Route
               path="/contact"
-              render={(props) => (
-                <Contact
-                  {...props}
-                  theme={propss.theme}
-                  setTheme={propss.setTheme}
-                />
-              )}
+              element={
+                <Contact theme={propss.theme} setTheme={propss.setTheme} />
+              }
             />
             <Route
               path="/splash"
-              render={(props) => (
-                <Splash
-                  {...props}
-                  theme={propss.theme}
-                  setTheme={propss.setTheme}
-                />
-              )}
+              element={
+                <Splash theme={propss.theme} setTheme={propss.setTheme} />
+              }
             />
             <Route
               path="/projects"
-              render={(props) => (
-                <Projects
-                  {...props}
-                  theme={propss.theme}
-                  setTheme={propss.setTheme}
-                />
-              )}
+              element={
+                <Projects theme={propss.theme} setTheme={propss.setTheme} />
+              }
             />
-          </Switch>
+          </Routes>
         </HashRouter>
       </div>
     );
@@ -94,70 +62,40 @@ export default function Main(propss) {
     return (
       <div>
         <HashRouter basename="/">
-          <Switch>
+          <Routes>
             <Route
               path="/"
-              exact
-              render={(props) => (
-                <Home
-                  {...props}
-                  theme={propss.theme}
-                  setTheme={propss.setTheme}
-                />
-              )}
+              element={<Home theme={propss.theme} setTheme={propss.setTheme} />}
             />
             <Route
               path="/home"
-              render={(props) => (
-                <Home
-                  {...props}
-                  theme={propss.theme}
-                  setTheme={propss.setTheme}
-                />
-              )}
+              element={<Home theme={propss.theme} setTheme={propss.setTheme} />}
             />
             <Route
               path="/experience"
-              exact
-              render={(props) => (
-                <Experience
-                  {...props}
-                  theme={propss.theme}
-                  setTheme={propss.setTheme}
-                />
-              )}
+              element={
+                <Experience theme={propss.theme} setTheme={propss.setTheme} />
+              }
             />
             <Route
               path="/education"
-              render={(props) => (
-                <Education
-                  {...props}
-                  theme={propss.theme}
-                  setTheme={propss.setTheme}
-                />
-              )}
+              element={
+                <Education theme={propss.theme} setTheme={propss.setTheme} />
+              }
             />
             <Route
               path="/contact"
-              render={(props) => (
-                <Contact
-                  {...props}
-                  theme={propss.theme}
-                  setTheme={propss.setTheme}
-                />
-              )}
+              element={
+                <Contact theme={propss.theme} setTheme={propss.setTheme} />
+              }
             />
             <Route
               path="/projects"
-              render={(props) => (
-                <Projects
-                  {...props}
-                  theme={propss.theme}
-                  setTheme={propss.setTheme}
-                />
-              )}
+              element={
+                <Projects theme={propss.theme} setTheme={propss.setTheme} />
+              }
             />
-          </Switch>
+          </Routes>
         </HashRouter>
       </div>
     );
